requirements.txt
aiofiles==23.2.1
markdown==3.5.2
weasyprint==59.0
pydyf==0.7.0
openai>=1.54.0
pydantic>=2.8.0
pydantic-core>=2.18.0
pydantic-settings>=2.4.0
pytest==8.0.0
pytest-asyncio==0.23.5
python-dotenv==1.0.1
langchain-community>=0.0.1 